requirements.txt
streamlit
openai
python-dotenv
requests
pandas
fpdf2
plotly